--- feature/client/src/desktopMain/kotlin/com/mifos/feature/client/clientList/ClientListScreen.desktop.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2025 Mifos Initiative
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- *
- * See https://github.com/openMF/android-client/blob/master/LICENSE.md
- */
-package com.mifos.feature.client.clientList
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.paging.PagingData
-import com.mifos.room.entities.client.ClientEntity
-import kotlinx.coroutines.flow.Flow
-
-@Composable
-actual fun LazyColumnForClientListApi(
-    pagingFlow: Flow<PagingData<ClientEntity>>,
-    isInSelectionMode: Boolean,
-    selectedItems: ClientSelectionState,
-    failedRefresh: () -> Unit,
-    onClientSelect: (Int) -> Unit,
-    selectedMode: () -> Unit,
-) {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            text = "Not available in desktop yet",
-        )
-    }
-}
